Fix CI
Some module dependencies have been updated and now require more facts.
Instead of managing them by hand, rely on rspec for that, but the facts
for some EOL systems are missing some core facts needed, so get rid of
them.

This is a short-term solution for fixing CI.  More love in this area
would be welcome!